backend/requirements.txt
anyio==4.9.0
arrow==1.3.0
asgiref==3.8.1
bidict==0.23.1
binaryornot==0.4.4
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cookiecutter==2.6.0
Django==5.2.3
django-cors-headers==4.8.0
django-debug-toolbar==5.2.0
django-flags==5.0.14
django-structlog==10.0.0
django-filter==25.1
djangorestframework==3.16.0
unittest-xml-reporting==3.2.0
drf_spectacular==0.29.0
fhir.resources==8.1.0
fhir_core==1.1.3
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
psycopg==3.2.12
psycopg_binary==3.2.12
psycopg-pool==3.2.7
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-multipart==0.0.20
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
six==1.17.0
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.47.2
text-unidecode==1.3
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
urllib3==2.5.0
WTForms==3.1.2
